+++ src/localis/indexes/__init__.py
@@ -1,3 +1,5 @@
 from .lookup_index import LookupIndex
 from .filter_index import FilterIndex
 from .search_index import SearchIndex
+
+__all__ = ["LookupIndex", "FilterIndex", "SearchIndex"]